Move out ArrayBase constructors to their own file
